refactor(subjects): move Subjects to src/
Closes #1555.
