+++ src/Rx.ts
@@ -120,9 +120,9 @@ export {Operator} from './Operator';
 export {Observer} from './Observer';
 export {Subscription, UnsubscriptionError} from './Subscription';
 export {Subscriber} from './Subscriber';
-export {AsyncSubject} from './subject/AsyncSubject';
-export {ReplaySubject} from './subject/ReplaySubject';
-export {BehaviorSubject} from './subject/BehaviorSubject';
+export {AsyncSubject} from './AsyncSubject';
+export {ReplaySubject} from './ReplaySubject';
+export {BehaviorSubject} from './BehaviorSubject';
 export {ConnectableObservable} from './observable/ConnectableObservable';
 export {Notification} from './Notification';
 export {EmptyError} from './util/EmptyError';
