+++ src/SubjectSubscription.ts
@@ -1,6 +1,6 @@
-import {Subject} from '../Subject';
-import {Observer} from '../Observer';
-import {Subscription} from '../Subscription';
+import {Subject} from './Subject';
+import {Observer} from './Observer';
+import {Subscription} from './Subscription';
 
 export class SubjectSubscription extends Subscription {
   isUnsubscribed: boolean = false;
