+++ src/operator/publishBehavior.ts
@@ -1,4 +1,4 @@
-import {BehaviorSubject} from '../subject/BehaviorSubject';
+import {BehaviorSubject} from '../BehaviorSubject';
 import {multicast} from './multicast';
 import {ConnectableObservable} from '../observable/ConnectableObservable';
